DEV-775: Add position sticky to nft details header
